Make failure condition of constraint consistent
Currently the constraint expects the property fn to return strictly
boolean true. The \QuickCheck\Property::shrinkLoop method on the other
hand allows any truthy value except exceptions (which is closer to
clojure.test.check behavior).

This commit makes the behavior of the PHPUnit constraint consistent with
the check in \QuickCheck\Property.